refactor: renames "index.ts" to "exports.ts" in "features/TextToImage/config/types"
